Add full page implementation
This is a refactoring to begin support for more of the Razor surface area
including @inject, Layouts and TagHelpers.

This change brings in the implementation of all of the required RazorPage
surface area based on the 1.0.0 RazorPage.

Also added a factory/activator abstraction and basic support for @inject.
Note that @injected things are not yet contextualized.