Adds support for @page
This change adds support for custom routing with pages. You can specify a
route with `@page "foo/bar"`. The quotes are required.

The route that you specify with `@page` will be concatenated to the path
to the page. So adding `@page "{id}"` to `Pages/Customer.razor` will match
the URL path `Customer/17`.

If you want a custom route from the root, use `~/` or `/` at the
beginning.